Move PathHelper inside StorageBlobClient
